+++ owGameM2/M2_Part_TextureTransform.h
@@ -7,7 +7,7 @@ class CM2_Part_TextureTransform
 public:
 	CM2_Part_TextureTransform(IFile* f, const SM2_TextureTransform& _proto, cGlobalLoopSeq global);
 
-	void calc(uint32 anim, uint32 time, uint32 globalTime);
+	void calc(uint16 anim, uint32 time, uint32 globalTime);
 
 	cmat4 getValue() const { return matrix; }
